Finished merge and restructured files
